@@ -24,72 +24,93 @@
     "standard-fix": "standard --fix ./components/*.js"
   },
   "dependencies": {
-    "apollo-fetch": "^0.6.0",
-    "apollo-link": "^0.5.4",
-    "chalk": "^2.3.0",
-    "compression": "^1.7.1",
+    "@zeit/next-sass": "^0.1.1",
+    "apollo-cache-inmemory": "^1.1.9",
+    "apollo-client": "^2.2.5",
+    "apollo-fetch": "^0.7.0",
+    "apollo-link": "^1.2.1",
+    "apollo-link-batch-http": "^1.2.0",
+    "apollo-link-error": "^1.0.7",
+    "apollo-link-http": "^1.5.2",
+    "apollo-link-persisted-queries": "^0.2.0",
+    "apollo-link-retry": "^2.2.2",
+    "apollo-link-token-refresh-before-refetch": "^0.1.1",
+    "apollo-logger": "^0.1.1",
+    "chalk": "^2.3.2",
+    "compression": "^1.7.2",
     "cors": "2.8.4",
-    "cross-env": "^5.1.1",
-    "cross-fetch": "^1.1.1",
-    "draft-js": "^0.10.4",
+    "cross-env": "^5.1.3",
+    "cross-fetch": "^2.0.0",
+    "draft-js": "^0.10.5",
     "draft-js-counter-plugin": "^2.0.1",
     "draft-js-emoji-plugin": "^2.0.1",
     "draft-js-hashtag-plugin": "^2.0.1",
-    "draft-js-plugins-editor": "^2.0.3",
-    "draftjs-to-html": "^0.8.1",
+    "draft-js-plugins-editor": "^2.0.4",
+    "draftjs-to-html": "^0.8.3",
     "express": "4.16.2",
-    "graphql": "^0.11.7",
-    "graphql-tag": "^2.6.0",
-    "helmet": "^3.9.0",
+    "graphql": "^0.13.1",
+    "graphql-tag": "^2.8.0",
+    "helmet": "^3.12.0",
     "immutable": "^3.8.2",
     "ip": "1.1.5",
-    "isomorphic-fetch": "2.2.1",
     "js-cookie": "^2.2.0",
+    "jwt-decode": "^2.2.0",
     "lru-cache": "4.1.1",
-    "moment": "^2.19.4",
-    "next": "4.2.1",
+    "moment": "^2.21.0",
+    "next": "5.0.0",
     "next-cookies": "^1.0.2",
-    "next-routes": "1.2.0",
+    "next-routes": "1.3.0",
     "node-sass": "^4.7.2",
-    "offline-plugin": "^4.8.5",
-    "pm2": "^2.8.0",
-    "rc-time-picker": "^3.2.1",
+    "offline-plugin": "^4.9.0",
+    "photoeditorsdk": "^4.3.1",
+    "pm2": "^2.10.1",
+    "rc-time-picker": "^3.3.1",
     "react": "^16.2.0",
-    "react-apollo": "^1.4.16",
-    "react-circular-progressbar": "^0.5.0",
-    "react-color": "^2.13.8",
+    "react-apollo": "^2.0.0-rc.0",
+    "react-circular-progressbar": "^0.8.0",
+    "react-color": "^2.14.0",
     "react-dom": "^16.2.0",
     "react-drag-and-drop": "^2.4.0",
-    "react-dropzone": "^4.2.3",
+    "react-dropzone": "^4.2.8",
     "react-gravatar": "^2.6.3",
     "react-helmet": "5.2.0",
+    "react-hot-loader": "^4.0.0",
     "react-lazyload": "^2.3.0",
-    "react-masonry-infinite": "^1.2.1",
+    "react-masonry-infinite": "^1.2.2",
+    "react-modal": "^3.3.2",
     "react-notify-toast": "^0.4.0",
-    "react-scroll": "^1.7.3",
+    "react-onclickoutside": "^6.7.1",
+    "react-scroll": "^1.7.7",
     "react-search-input": "^0.11.3",
-    "react-tag-input": "^4.7.2",
+    "react-tag-input": "^4.8.2",
     "redux": "3.7.2",
-    "semantic-ui-react": "^0.77.1",
-    "twitter-text": "^1.14.7"
+    "semantic-ui-react": "^0.78.3",
+    "symlink-dir": "^1.1.2",
+    "twitter-text": "^2.0.4"
   },
   "devDependencies": {
-    "babel-eslint": "^8.0.3",
+    "autoprefixer": "^8.1.0",
+    "babel-eslint": "^8.2.2",
+    "babel-loader": "^7.1.3",
+    "babel-plugin-inline-dotenv": "^1.1.2",
+    "babel-plugin-inline-import-graphql-ast": "^2.2.0",
     "babel-plugin-inline-react-svg": "^0.5.2",
-    "babel-plugin-inline-dotenv": "^1.1.1",
-    "babel-plugin-inline-import-graphql-ast": "^2.0.4",
-    "babel-plugin-transform-assets-import-to-string": "^1.0.1",
-    "babel-plugin-transform-inline-environment-variables": "^0.2.0",
-    "css-loader": "^0.28.7",
-    "lint-staged": "6.0.0",
-    "ngrok": "2.2.24",
-    "prettier": "1.9.1",
-    "release": "3.0.2",
-    "sass-loader": "^6.0.6",
-    "standard": "^10.0.3",
-    "style-loader": "^0.19.0",
-    "webpack": "^3.10.0",
-    "webpack-bundle-analyzer": "2.9.1"
+    "babel-plugin-transform-assets-import-to-string": "^1.1.0",
+    "babel-plugin-transform-inline-environment-variables": "^0.3.0",
+    "css-loader": "^0.28.11",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "lint-staged": "7.0.0",
+    "ngrok": "3.0.0",
+    "postcss-loader": "^2.1.1",
+    "prettier": "1.11.1",
+    "raw-loader": "^0.5.1",
+    "release": "3.1.6",
+    "sass-loader": "^6.0.7",
+    "standard": "^11.0.0",
+    "style-loader": "^0.20.2",
+    "url-loader": "^1.0.1",
+    "webpack": "^3.11.0",
+    "webpack-bundle-analyzer": "2.11.1"
   },
   "engines": {
     "node": ">=8.1.0",
