@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
 import { Icon } from 'semantic-ui-react'
-import { EditorState, ContentState } from 'draft-js'
 import Editor, { createEditorStateWithText } from 'draft-js-plugins-editor'
 
 import createHashtagPlugin from 'draft-js-hashtag-plugin'
@@ -8,11 +7,9 @@ import createEmojiPlugin from 'draft-js-emoji-plugin'
 import createCounterPlugin from 'draft-js-counter-plugin'
 import twitter from 'twitter-text'
 
-if (process.env.BROWSER) {
-  require('./style.scss')
-  require('./draft-emoji.scss')
-  require('./draft-counter.scss')
-}
+import './style.scss'
+import './draft-emoji.scss'
+import './draft-counter.scss'
 
 export default class TextEditor extends Component {
   constructor (props) {
@@ -25,15 +22,17 @@ export default class TextEditor extends Component {
     this._counterPlugin = createCounterPlugin()
   }
 
-  componentWillReceiveProps (nextProps) {
-    if (this.props.defaultValue !== nextProps.defaultValue) {
-      // console.log('WillReceive props')
-      // this.updateEditorState(nextProps.defaultValue)
+  componentDidMount () {
+    if (this.props.editEnabled) {
+      this.editor.focus()
     }
   }
 
   focus = (e) => {
-    this.editor.focus()
+    console.log(this.editor)
+    if (!this.editor.isFocused) {
+      this.editor.focus()
+    }
   }
 
   onChange = (editorState) => {
@@ -53,6 +52,18 @@ export default class TextEditor extends Component {
     return str ? twitter.getTweetLength(str) : 0
   }
 
+  getLimit = () => {
+    const { socialNetwork } = this.props
+    switch (socialNetwork) {
+      case 'TWITTER':
+        return 280
+      case 'INSTAGRAM':
+        return 2200
+      case 'FACEBOOK':
+        return
+    }
+  }
+
   render () {
     const { EmojiSuggestions, EmojiSelect } = this._emojiPlugin
     const { CustomCounter } = this._counterPlugin
@@ -64,7 +75,7 @@ export default class TextEditor extends Component {
     ]
 
     return (<div key={this.props.index} id={this.props.id} className='content-editor'>
-      <div className='content-editor-wrapper' onClick={this.focus}>
+      <div className='content-editor-wrapper'>
         <Editor
           editorState={this.state.editorState}
           onChange={this.onChange}
@@ -78,7 +89,7 @@ export default class TextEditor extends Component {
         {this.props.attachMediaIcon && <div className='attach-media'><Icon name='image' onClick={this.props.onShowAttachAssets} /></div>}
       </div>
       <div className='emoji-picker'>
-        <CustomCounter countFunction={this.customCountFunction} limit={140} />
+        <CustomCounter countFunction={this.customCountFunction} limit={this.getLimit()}/>
         <EmojiSelect />
       </div>
     </div>)
