@@ -1,91 +1,36 @@
 export default () => [
   <link
-    key='0'
-    rel='apple-touch-icon'
-    sizes='57x57'
-    href='/static/icons/apple-icon-57x57.png'
-  />,
-  <link
-    key='1'
-    rel='apple-touch-icon'
-    sizes='60x60'
-    href='/static/icons/apple-icon-60x60.png'
-  />,
-  <link
-    key='2'
-    rel='apple-touch-icon'
-    sizes='72x72'
-    href='/static/icons/apple-icon-72x72.png'
-  />,
-  <link
-    key='3'
-    rel='apple-touch-icon'
-    sizes='76x76'
-    href='/static/icons/apple-icon-76x76.png'
-  />,
-  <link
-    key='4'
-    rel='apple-touch-icon'
-    sizes='114x114'
-    href='/static/icons/apple-icon-114x114.png'
-  />,
-  <link
-    key='15'
-    rel='apple-touch-icon'
-    sizes='120x120'
-    href='/static/icons/apple-icon-120x120.png'
-  />,
-  <link
-    key='5'
-    rel='apple-touch-icon'
-    sizes='144x144'
-    href='/static/icons/apple-icon-144x144.png'
-  />,
-  <link
-    key='6'
-    rel='apple-touch-icon'
-    sizes='152x152'
-    href='/static/icons/apple-icon-152x152.png'
-  />,
-  <link
-    key='7'
     rel='apple-touch-icon'
     sizes='180x180'
-    href='/static/icons/apple-icon-180x180.png'
+    href='/static/icons/apple-touch-icon.png?v=2'
   />,
   <link
-    key='8'
-    rel='icon'
-    type='image/png'
-    sizes='192x192'
-    href='/static/icons/android-icon-192x192.png'
-  />,
-  <link
-    key='9'
     rel='icon'
     type='image/png'
     sizes='32x32'
-    href='/static/icons/favicon-32x32.png'
+    href='/static/icons/favicon-32x32.png?v=2'
   />,
   <link
-    key='10'
-    rel='icon'
-    type='image/png'
-    sizes='96x96'
-    href='/static/icons/favicon-96x96.png'
-  />,
-  <link
-    key='11'
     rel='icon'
     type='image/png'
     sizes='16x16'
-    href='/static/icons/favicon-16x16.png'
+    href='/static/icons/favicon-16x16.png?v=2'
   />,
-  <meta key='12' name='msapplication-TileColor' content='#f9f9f9' />,
+  <link
+    rel='manifest'
+    href='/static/icons/site.webmanifest?v=2'
+  />,
+  <link
+    rel='mask-icon'
+    href='/static/icons/safari-pinned-tab.svg?v=2'
+    color='#5bbad5'
+  />,
   <meta
-    key='13'
-    name='msapplication-TileImage'
-    content='/static/icons/ms-icon-144x144.png'
+    name='msapplication-TileColor'
+    content='#da532c'
   />,
-  <meta key='14' name='theme-color' content='#f9f9f9' />
+  <meta
+    name='theme-color'
+    content='#323945'
+  />
 ]
