@@ -1,8 +1,8 @@
 import gql from 'graphql-tag'
 
 export const getAssets = gql`
-query getAssets ($categoryIds: [ID], $after: String) {
-  media_find (categoryIds: $categoryIds, first: 10, after: $after) {
+query getAssets ($categoryIds: [ID]) {
+  media_find (categoryIds: $categoryIds, inBank: true) {
     pageInfo {
       hasNextPage
       endCursor
@@ -12,6 +12,7 @@ query getAssets ($categoryIds: [ID], $after: String) {
       node {
         id
         url
+        mime
         categories {
           id
           name
