@@ -5,38 +5,42 @@ import {getAssets} from './graphql/AssetQueries'
 import Loader from '../Loader'
 import AssetDrop from './AssetDrop'
 
-if (process.env.BROWSER) {
-  require('./dropdown.scss')
-}
+import './dropdown.scss'
 
-const AssetDropdown = ({socialId, uploadAsset, onSelectAsset, assetData}) => {
-  if (assetData.loading) {
-    return <Loader />
-  }
+class AssetDropdown extends React.Component {
+  render() {
+    const {socialId, uploadAsset, onSelectAsset, assetData, categoryIds, categories} = this.props
+    if (assetData.loading) {
+      return <Loader/>
+    }
 
-  if (assetData.error) {
-    return (<div className='page-wrapper'>
-      {assetData.error.message}
-    </div>)
+    if (assetData.error) {
+      return (<div className='page-wrapper'>
+        {assetData.error.message}
+      </div>)
+    }
+    const assets = assetData.media_find.edges
+    let tags = new Map()
+    categories.map(category => tags.set(category.id, category))
+    return (
+      <AssetDrop
+        assets={assets}
+        tags={tags}
+        socialId={socialId}
+        onSelectAsset={onSelectAsset}
+        categoryIds={categoryIds}
+        simpleUpload
+        onUpload={this.props.onUpload}
+      />
+    )
   }
-  const assets = assetData.media_find.edges
-  const tags = []
-  assets.map(asset => asset.node.categories.map(cat => tags.push(cat)))
-  return (
-    <AssetDrop
-      assets={assets}
-      tags={tags}
-      socialId={socialId}
-      onSelectAsset={onSelectAsset}
-    />
-  )
 }
 
 export default compose(
   graphql(getAssets, {
     options: (ownProps) => ({
       variables: {
-        categoryIds: []
+        categoryIds: ownProps.categories.map(cat => cat.id)
       }
     }),
     name: 'assetData'
