@@ -1,19 +1,30 @@
 import React, { Component } from 'react'
 import { graphql, compose } from 'react-apollo'
+import gql from 'graphql-tag'
 import Notification from '../Notification'
-import { Icon, Table, Image, Popup, Input } from 'semantic-ui-react'
+import ColorPickerCell from './ColorPickerCell'
+import ColorNames from './consts/color.json'
+import persist from '../../libraries/persist'
+import { Icon, Table, Popup, Button, Input, Modal, Confirm, Image } from 'semantic-ui-react'
 import TimePicker from 'rc-time-picker'
 import moment from 'moment'
 
 // import Freshness from '../Freshness'
 
-import { isEmpty, getCorrectHours } from '../../libraries/helpers'
+import Sun from '../../static/images/sun.svg'
+import Afternoon from '../../static/images/afternoon.svg'
+import Morning from '../../static/images/morning.svg'
+import Night from '../../static/images/night.svg'
+
+import {isEmpty, getCorrectHours, hexToRGB, calendarTimeAndDayWithOffset} from '../../libraries/helpers'
 
 import { createCalendarSlotMutation, updateCalendarSlotMutation, deleteCalendarSlotMutation, editCalendarRowTimeMutation } from './graphql/calendarMutations'
+import { createCategoryMutation } from './graphql/categoryMutations'
+import { getCategorybySocialIdQuery } from './graphql/categoryQueries'
+import { calendarSlotsFindQuery } from './graphql/contentQueries';
+import calendarSlotsGql from '../Posts/graphql/calendarSlots.gql'
 
-if (process.env.BROWSER) {
-  require('./styles/rc-time-picker.scss')
-}
+import './styles/rc-time-picker.scss'
 
 const format = 'HH:mm'
 
@@ -22,7 +33,12 @@ class CalendarCell extends Component {
     super(props)
     this.state = {
       isShowPopup: false,
-      categories: this.props.categories
+      categories: this.props.categories,
+      categoryName: '',
+      categoryColor: '#36BF99',
+      cellTime: this.props.cellTime,
+      timePickerOpen: false,
+      confirmOpen: false
     }
   }
 
@@ -46,7 +62,8 @@ class CalendarCell extends Component {
     const socialProfileId = this.props.socialId
     const calendarSlotType = 'WEEKLY'
     const categoryId = category.id
-    this.props.createCalendarSlot({ calendarSlotDay, calendarSlotTime, calendarSlotType, categoryId, socialProfileId })
+    const {day, time} = calendarTimeAndDayWithOffset({day: calendarSlotDay, time: calendarSlotTime}, - new Date().getTimezoneOffset())
+    this.props.createCalendarSlot(day, time, calendarSlotType, categoryId, socialProfileId)
     .then((data) => {
       Notification.success('Create Calendarslot is success.')
     })
@@ -55,30 +72,41 @@ class CalendarCell extends Component {
   onEditCalendarSlot = (calendarSlotDay, calendarSlotTime, calendarSlotId, category) => {
     const calendarSlotType = 'WEEKLY'
     const categoryId = category.id
-    this.props.editCalendarSlot({calendarSlotId, calendarSlotDay, calendarSlotTime, calendarSlotType, categoryId})
+    const {day, time} = calendarTimeAndDayWithOffset({day: calendarSlotDay, time: calendarSlotTime}, - new Date().getTimezoneOffset())
+    this.props.editCalendarSlot(calendarSlotId, day, time, calendarSlotType, categoryId)
     .then((data) => {
       Notification.success('Edit Calendarslot is success.')
     })
   }
 
   onDelEvent = (time) => {
-    const deletedTime = this.getCalendarTime(time)
-    const {slot} = this.props
+    time = this.getCalendarTime(time)
+    const {slot, slots} = this.props
+    let newSlots = slots
+    delete newSlots[time]
+    this.setState({confirmOpen: false})
+    this.props.onSlotsUpdate(newSlots)
     Object.keys(slot).forEach((day) => {
-      this.props.onDelEvent(deletedTime)
       if (slot[day].id !== undefined) {
-        this.onDeleteCalendarSlot(slot[day].id)
+        this.props.deleteCalendarSlot({ id: slot[day].id })
+          .then((data) => {
+            Notification.success('Remove Calendar Slot is success.')
+          })
+          .catch((err) => {
+            console.log(err)
+            this.props.onSlotsUpdate(slots)
+          })
       }
     })
   }
 
   onDeleteCalendarSlot = (id) => {
-    this.props.deleteCalendarSlot({ id })
-    const cellTime = this.props.cellTime
+    const cellTime = this.state.cellTime
     const slots = this.props.slots
     const copySlots = slots
     slots[cellTime][this.props.day] = {}
     this.props.onSlotsUpdate(slots)
+    this.props.deleteCalendarSlot({ id })
     .then((data) => {
       Notification.success('Remove Calendar Slot is success.')
     })
@@ -89,37 +117,42 @@ class CalendarCell extends Component {
   }
 
   onTimeUpdate = (value) => {
+    if (value) {
+      this.setState({
+        cellTime: this.getCalendarTime(value)
+      })
+    }
+  }
+
+  onSaveRow = () => {
     const cellTime = this.props.cellTime
-    const newTime = this.getCalendarTime(value)
+    const newTime = this.state.cellTime
     const slots = this.props.slots
     const oldDays = slots[cellTime]
     const calendarSlotIds = []
 
-    Object.keys(oldDays).forEach((day) => {
-      if (oldDays[day].id) {
-        oldDays[day].time = newTime
-        calendarSlotIds.push(oldDays[day].id)
-      }
-    })
-    slots[newTime] = oldDays
-    delete slots[cellTime]
-    Object.keys(slots).sort()
+    this.setState({timePickerOpen: false})
+    if (cellTime !== newTime) {
+      Object.keys(oldDays).forEach((day) => {
+        if (oldDays[day].id) {
+          oldDays[day].time = newTime
+          calendarSlotIds.push(oldDays[day].id)
+        }
+      })
+      slots[newTime] = oldDays
+      delete slots[cellTime]
+      Object.keys(slots).sort()
 
-    this.props.onSlotsUpdate(slots)
-    this.props.editCalendarRowTime({
-      calendarSlotIds,
-      newTime
-    }).then((data) => {
-      Notification.success('Update Calendar Slot is success.')
-    })
-  }
+      this.props.onSlotsUpdate(slots)
 
-  hexToRGB = (hex, opacity) => {
-    hex = parseInt(hex.slice(1), 16)
-    let r = hex >> 16
-    let g = hex >> 8 & 0xFF
-    let b = hex & 0xFF
-    return `rgba(${r},${g},${b},${opacity})`
+      const {time} = calendarTimeAndDayWithOffset({time: newTime, day: 0}, -new Date().getTimezoneOffset())
+      this.props.editCalendarRowTime({
+        calendarSlotIds,
+        newTime: time
+      }).then((data) => {
+        Notification.success('Update Calendar Slot is success.')
+      })
+    }
   }
 
   handleOpen = () => {
@@ -130,38 +163,118 @@ class CalendarCell extends Component {
     this.setState({ isShowPopup: false })
   }
 
+  onCategoryCreate = () => {
+    let {searchTerm, isOpen, categoryColor} = this.state
+
+    const socialProfileId = persist.willGetCurrentSocialProfile()
+
+    if (searchTerm !== '') {
+      this.props.createCategory({categoryName: searchTerm, socialProfileId, color: categoryColor})
+      .then((data) => {
+        Notification.success('New Category is added.')
+        this.setState({
+          isOpen: !isOpen
+        })
+      })
+    }
+  }
+
+  onCategoryColorChange = (colorCode) => {
+    this.setState({
+      categoryColor: colorCode
+    })
+  }
+
+  confirmDeleteRow = () => {
+    this.setState({confirmOpen: true})
+  }
+
+  renderCreateNewCategory = () => {
+    return (<div className='add-category-item'>
+      <b>Category not found. Create a new one.</b>
+      <div className='dropdown-content'>
+        <div className='category-color-view'>
+          <p className='category-title'>
+            Select a Color
+          </p>
+
+          <div className='category-color-picker'>
+            {
+              ColorNames.color.map((color, index) => (
+                <ColorPickerCell
+                  key={index}
+                  colorCode={color.name}
+                  categoryColorChange={this.onCategoryColorChange}
+                />
+              ))
+            }
+          </div>
+        </div>
+        <div className='category-button-view'>
+          {/*<Button onClick={this.handleClose}>Cancel</Button>*/}
+          <Button className='savecategory-button' onClick={this.onCategoryCreate}>Create Category</Button>
+        </div>
+      </div>
+    </div>)
+  }
+
   render () {
-    let { cellTime, type, day, obj } = this.props
+    let { type, day, obj } = this.props
+    let { cellTime, timePickerOpen } = this.state
     const time = moment(getCorrectHours(cellTime), 'HH:mm')
     if (type === 'time') {
-      let sunIcon = '/static/images/sun.png'
-      if (cellTime < 720) { // Before mid-day
-        sunIcon = '/static/images/sunrise.png'
-      }
-      if (cellTime > 1080) { // After 6PM
-        sunIcon = '/static/images/sunset.png'
+      let sunIcon = null
+      switch (true) {
+        case cellTime < 240: // Before 4am:
+          sunIcon = <Night/>
+          break;
+        case cellTime < 720: // Before mid-day
+          sunIcon = <Morning/>
+          break;
+        case cellTime < 1020: // Before 5PM
+          sunIcon = <Sun/>
+          break;
+        case cellTime < 1200: // Before 8PM
+          sunIcon = <Afternoon/>
+          break;
+        default: // After 8PM
+          sunIcon = <Night/>
+          break;
       }
       return (
-        <Table.Cell className='calendar-time-cell'>
-          <Image inline src={sunIcon} />
+        <Table.Cell key={time} className='calendar-time-cell'>
+          {sunIcon}
           <TimePicker
             id='custom-time-picker'
             value={time}
             showSecond={false}
+            onOpen={() => this.setState({timePickerOpen: true})}
+            onClose={this.onSaveRow}
             onChange={this.onTimeUpdate}
             format={format}
+            open={timePickerOpen}
           />
-          <Icon name='trash' className='remove-icon-row' onClick={() => this.onDelEvent(time)} />
+          <Icon name='trash' className='remove-icon-row' onClick={() => this.confirmDeleteRow()} />
+          <Confirm
+            open={this.state.confirmOpen}
+            onCancel={() => this.setState({confirmOpen: false})}
+            onConfirm={() => this.onDelEvent(time)}
+            content={<div className='content'>Are you sure you want to delete this time?<br /><br />All calendar slots in this row will be removed.</div>}
+            size='tiny'
+          />
         </Table.Cell>
       )
     }
 
+    const filteredCategories = this.state.searchTerm ? this.state.categories.filter(item => item.node.name.toLowerCase().includes(this.state.searchTerm.toLowerCase())) : this.state.categories || []
+
     if (obj.category) {
-      const backColor = this.hexToRGB(obj.category.color, 0.25)
-      const hoverCatColor = this.hexToRGB(obj.category.color, 0.4)
-      const catColor = this.hexToRGB(obj.category.color, 1)
+      const backColor = hexToRGB(obj.category.color, 0.25)
+      const hoverCatColor = hexToRGB(obj.category.color, 0.4)
+      const catColor = hexToRGB(obj.category.color, 1)
       return (
         <Table.Cell
+          key={obj.id}
           className='calendar-cell no-border'
           style={{ backgroundColor: backColor }}
         >
@@ -177,6 +290,12 @@ class CalendarCell extends Component {
               {obj.category.name}
             </span>
           </div>
+          <Icon
+            name='trash'
+            className='remove-icon'
+            onClick={() => this.onDeleteCalendarSlot(obj.id)}
+            style={{color: catColor}}
+          />
           {/*<Popup
             trigger={<div
               className='category-label'
@@ -203,20 +322,55 @@ class CalendarCell extends Component {
             onClose={this.handleClose}
             position='bottom center'
           />*/}
-          <Icon
-            name='trash'
-            className='remove-icon'
-            onClick={() => this.onDeleteCalendarSlot(obj.id)}
-            style={{color: catColor}}
-          />
+          {this.props.mobile &&
+            <Popup
+              className='mobile-menu'
+              trigger={<Icon name='ellipsis vertical' style={{ backgroundColor: this.props.catId === obj.category.id ? hoverCatColor : 'transparent' }} />}
+              content={<div className='mobile-menu-inner'>
+                <div className='item' onClick={() => this.setState({showCategoryModal: true})}>Replace</div>
+                <div className='item' onClick={() => this.onDeleteCalendarSlot(obj.id)}>
+                  Clear
+                </div>
+              </div>
+              }
+              on='click'
+              position='bottom right'
+            />}
+          {this.props.mobile && this.state.showCategoryModal && <Modal
+            open={this.state.showCategoryModal}
+            onClose={() => this.setState({showCategoryModal: false})}
+            className='post-item-pane'
+            size='small'
+          >
+            <Modal.Content scrolling>
+              <div className='add-category-popup'>
+                <Input
+                  placeholder='Search...'
+                  onChange={(event, {value}) => this.setState({searchTerm: value})}
+                />
+                {filteredCategories.length > 0 ?
+                  <ul className='add-category-list'>
+                    {filteredCategories.map(cat => {
+                      return (
+                        <li key={cat.node.name} className='add-category-item' onClick={() => this.onCellUpdate(day, time, obj, cat)}>
+                          <div className='add-category-color' style={{backgroundColor: cat.node.color}} />
+                          {cat.node.name}
+                        </li>
+                      )
+                    })}
+                  </ul>
+                  :
+                  this.renderCreateNewCategory()
+                }
+              </div>
+            </Modal.Content>
+          </Modal>}
         </Table.Cell>
       )
     }
 
-    const filteredCategories = this.state.searchTerm ? this.state.categories.filter(item => item.node.name.includes(this.state.searchTerm)) : this.state.categories || []
-
     return (
-      <Table.Cell className='calendar-cell'>
+      <Table.Cell key={obj.id} className='calendar-cell'>
         <Popup
           trigger={<div
             className='addcategory-label'
@@ -229,22 +383,20 @@ class CalendarCell extends Component {
               placeholder='Search...'
               onChange={(event, {value}) => this.setState({searchTerm: value})}
             />
-            <ul className='add-category-list'>
-              {filteredCategories.length > 0 ?
-                filteredCategories.map(cat => {
+            {filteredCategories.length > 0 ?
+              <ul className='add-category-list'>
+                {filteredCategories.map(cat => {
                   return (
-                    <li className='add-category-item' onClick={() => this.onCellUpdate(day, time, obj, cat)}>
+                    <li key={cat.node.name} className='add-category-item' onClick={() => this.onCellUpdate(day, time, obj, cat)}>
                       <div className='add-category-color' style={{backgroundColor: cat.node.color}} />
                       {cat.node.name}
                     </li>
                   )
-                })
-                :
-                <li className='add-category-item'>
-                  Categories not found
-                </li>
-              }
-            </ul>
+                })}
+              </ul>
+              :
+              this.renderCreateNewCategory()
+            }
           </div>}
           on='click'
           open={this.state.isShowPopup}
@@ -260,7 +412,7 @@ export default compose(
   graphql(createCalendarSlotMutation, {
     props ({ ownProps, mutate }) {
       return {
-        createCalendarSlot ({ calendarSlotDay, calendarSlotTime, calendarSlotType, categoryId, socialProfileId }) {
+        createCalendarSlot (calendarSlotDay, calendarSlotTime, calendarSlotType, categoryId, socialProfileId) {
           return mutate({
             variables: {
               input: {
@@ -270,6 +422,19 @@ export default compose(
                 categoryId,
                 socialProfileId
               }
+            },
+            update: (proxy, {data: {createCalendarSlot}}) => {
+              // Posts page query
+              let postPageData = proxy.readQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [socialProfileId], type: 'WEEKLY'}})
+              let calendarSlot = createCalendarSlot.calendarSlot
+              delete calendarSlot.contentSchedules
+              postPageData.calendarSlots_find.edges.push({node: calendarSlot, __typename: 'CalendarSlotsEdge'})
+              proxy.writeQuery({ query: gql`${calendarSlotsGql}`, variables: {profileIds: [socialProfileId], type: 'WEEKLY'}, data: postPageData })
+
+              // Calendar page query
+              let calendarPageData = proxy.readQuery({query: calendarSlotsFindQuery, variables: {profileIds: [socialProfileId], type: 'WEEKLY'}})
+              calendarPageData.calendarSlots_find.edges.push({node: createCalendarSlot.calendarSlot, __typename: 'CalendarSlotsEdge'})
+              proxy.writeQuery({ query: calendarSlotsFindQuery, variables: {profileIds: [socialProfileId], type: 'WEEKLY'}, data: calendarPageData })
             }
           })
         }
@@ -279,7 +444,7 @@ export default compose(
   graphql(updateCalendarSlotMutation, {
     props ({ ownProps, mutate }) {
       return {
-        editCalendarSlot ({ calendarSlotId, calendarSlotDay, calendarSlotTime, calendarSlotType, categoryId }) {
+        editCalendarSlot (calendarSlotId, calendarSlotDay, calendarSlotTime, calendarSlotType, categoryId) {
           return mutate({
             variables: {
               input: {
@@ -288,6 +453,29 @@ export default compose(
                 calendarSlotTime,
                 calendarSlotType,
                 categoryId
+              }
+            },
+            update: (proxy, {data: {editCalendarSlot}}) => {
+              // Posts page query
+              let postPageData = proxy.readQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}})
+              let calendarSlot = createCalendarSlot.calendarSlot
+              delete calendarSlot.contentSchedules
+              for (const [index, calendarSlot] of postPageData.calendarSlots_find.edges.entries()) {
+                if (calendarSlot.node.id === editCalendarSlot.calendarSlot.id) {
+                  postPageData.calendarSlots_find.edges[index].node = editCalendarSlot.calendarSlot
+                  proxy.writeQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}, data: postPageData})
+                  break
+                }
+              }
+
+              // Calendar page query
+              let calendarPageData = proxy.readQuery({query: calendarSlotsFindQuery, variables: {profileIds: [socialProfileId], type: 'WEEKLY'}})
+              for (const [index, calendarSlot] of calendarPageData.calendarSlots_find.edges.entries()) {
+                if (calendarSlot.node.id === editCalendarSlot.calendarSlot.id) {
+                  calendarPageData.calendarSlots_find.edges[index].node = editCalendarSlot.calendarSlot
+                  proxy.writeQuery({query: calendarSlotsFindQuery, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}, data: calendarPageData})
+                  break
+                }
               }
             }
           })
@@ -303,6 +491,27 @@ export default compose(
             variables: {
               input: {
                 id
+              }
+            },
+            update: (proxy, {data: {deleteCalendarSlot}}) => {
+              // Posts page query
+              let postPageData = proxy.readQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}})
+              for (const [index, calendarSlot] of postPageData.calendarSlots_find.edges.entries()) {
+                if (calendarSlot.node.id === id) {
+                  postPageData.calendarSlots_find.edges.splice(index, 1)
+                  proxy.writeQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}, data: postPageData})
+                  break
+                }
+              }
+
+              // Calendar page query
+              let calendarPageData = proxy.readQuery({query: calendarSlotsFindQuery, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}})
+              for (const [index, calendarSlot] of calendarPageData.calendarSlots_find.edges.entries()) {
+                if (calendarSlot.node.id === id) {
+                  calendarPageData.calendarSlots_find.edges.splice(index , 1)
+                  proxy.writeQuery({query: calendarSlotsFindQuery, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}, data: calendarPageData})
+                  break
+                }
               }
             }
           })
@@ -320,7 +529,53 @@ export default compose(
                 calendarSlotIds,
                 newTime
               }
+            },
+            update: (proxy, {data: {editCalendarRowTime}}) => {
+              // Posts page query
+              let postPageData = proxy.readQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}})
+              for (const [index, calendarSlot] of postPageData.calendarSlots_find.edges.entries()) {
+                calendarSlotIds.forEach((id) => {
+                  if (calendarSlot.node.id === id) {
+                    postPageData.calendarSlots_find.edges[index].node.time = newTime
+                  }
+                })
+              }
+              proxy.writeQuery({query: gql`${calendarSlotsGql}`, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}, data: postPageData})
+
+              // Calendar page query
+              let calendarPageData = proxy.readQuery({query: calendarSlotsFindQuery, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}})
+              for (const [index, calendarSlot] of calendarPageData.calendarSlots_find.edges.entries()) {
+                calendarSlotIds.forEach((id) => {
+                  if (calendarSlot.node.id === id) {
+                    calendarPageData.calendarSlots_find.edges[index].node.time = newTime
+                  }
+                })
+                proxy.writeQuery({query: calendarSlotsFindQuery, variables: {profileIds: [ownProps.socialId], type: 'WEEKLY'}, data: calendarPageData})
+              }
             }
+          })
+        }
+      }
+    }
+  }),
+  graphql(createCategoryMutation, {
+    props ({ownProps, mutate}) {
+      return {
+        createCategory ({categoryName, socialProfileId, color}) {
+          return mutate({
+            variables: {
+              input: {
+                categoryName,
+                socialProfileId,
+                color
+              }
+            },
+            refetchQueries: [{
+              query: getCategorybySocialIdQuery,
+              variables: {
+                id: ownProps.socialId
+              }
+            }]
           })
         }
       }
