@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { Icon } from 'semantic-ui-react'
+import { Icon, Image, Popup } from 'semantic-ui-react'
 import { withApollo } from 'react-apollo'
 import { Draggable } from 'react-drag-and-drop'
 import MasonryInfiniteScroller from 'react-masonry-infinite'
@@ -8,9 +8,7 @@ import Notification from '../Notification'
 import Loader from '../Loader'
 import {fetchOneCategoryQuery} from './graphql/categoryQueries'
 
-if (process.env.BROWSER) {
-  require('./styles/react-search-input.scss')
-}
+import './styles/react-search-input.scss'
 
 class CampaignsItems extends Component {
   constructor (props) {
@@ -35,7 +33,7 @@ class CampaignsItems extends Component {
       loading: true
     })
     this.props.client.query({
-      query: fetchOneCategoryQuery, 
+      query: fetchOneCategoryQuery,
       variables: {
         id: this.props.categoryId,
         after: endCursor
@@ -61,6 +59,28 @@ class CampaignsItems extends Component {
     })
   }
 
+  renderImages = (media) => {
+    return media.map(m => <Image className='post-image' src={m.url} inline/>)
+  }
+
+  countLikeFavsPosted = (schedules) => {
+    let list = schedules.edges.map((m) => m.node);
+    let likes = 0;
+    let favs = 0;
+    let posted = list.length;
+    if(list.length > 0) {
+      list.map(({status}) => {
+        if(status && status.favouriteCount){
+          likes = parseInt(likes) + parseInt(status.favouriteCount)
+        }
+        if(status && status.retweetCount){
+          favs = parseInt(favs) + parseInt(status.retweetCount)
+        }
+      })
+    }
+    return {likes, favs, posted};
+  }
+
   render () {
     const {items} = this.state
 
@@ -80,30 +100,45 @@ class CampaignsItems extends Component {
       packed='data-packed'
       hasMore={this.state.hasMore}
     >*/}
-      {items.map((content, index) =>
-        <div key={content.id} className='post-short-item'>
-          <div className='post-short-content'>
-            <div className='post-short-stat'>
-              <div>
-                <Icon name='twitter' />
-                <Icon name='facebook' />
-                <Icon name='instagram' />
+      {items.map((content, index) => {
+        let {likes, favs, posted} = this.countLikeFavsPosted(content.schedules);
+        return (<Draggable type='post' data={JSON.stringify(content)}>
+          <div key={content.id} className='post-short-item'>
+            <div className='post-short-content'>
+              <div className='post-short-stat'>
+                <Popup
+                  key={0}
+                  trigger={<Icon name='exchange' />}
+                  content='Drag and Drop into your post'
+                  inverted
+                />
+                <div>
+                  <Icon name='twitter' />
+                  <Icon name='facebook' />
+                  <Icon name='instagram' />
+                </div>
+                <div className='social-stat'>
+                  <span><Icon name='heart' />{likes}</span>
+                  <span><Icon name='refresh' />{favs}</span>
+                  {/* <span><Icon name='comment' />10</span> */}
+                </div>
+                <div>
+                  <b>Times posted:&nbsp;{posted}</b>
+                </div>
               </div>
-              <div className='social-stat'>
-                <span><Icon name='heart' />10</span>
-                <span><Icon name='refresh' />10</span>
-                <span><Icon name='comment' />10</span>
+              {content.text && <div className='post-short-desc'>
+                {content.text}
               </div>
-              <div>
-                <b>Times posted:&nbsp;2</b>
-              </div>
-            </div>
-            <div className='post-short-desc'>
-              {content.text}
+              }
+              {!!content.media.length &&
+                this.renderImages(content.media)
+              }
             </div>
           </div>
-        </div>
-      )}
+        </Draggable>)
+      })}
+      {this.props.mobile && <div>
+      </div>}
     </div>)
   }
 }
