@@ -1,9 +1,10 @@
-import React from 'react'
-import {Tab, Icon, Button, Popup, Label, Image} from 'semantic-ui-react'
-import Logo from '../../../static/images/proofer-logo.svg'
+  import React, {Component} from 'react'
+import {Button, Icon, Image, Popup, Tab} from 'semantic-ui-react';
 import ContentComments from '../../Comments'
+import ImageGrid from "../../ImageGrid/index";
+  import {hexToRGB} from "../../../libraries/helpers";
 
-class NormalState extends React.Component {
+export default class NormalState extends Component {
   constructor (props) {
     super(props)
     this.state = {
@@ -13,65 +14,125 @@ class NormalState extends React.Component {
 
   onToggleComment = () => this.setState({showComments: !this.state.showComments})
 
-  renderMedia = (media, index) => {
-    return (<div key={index} className='content-media'>
-      {media.type === 'image' ?
-        <Image src={media.url} className='pull-right custom-image-style' />
-        :
+  renderAllMedia (assets) {
+    const imageMimes = ['image/jpeg', 'image/png', 'image/gif']
+    const videoMimes = ['video/quicktime', 'video/mp4']
+    let videoIndex = false
+
+    for (let media of assets) {
+      if (videoMimes.indexOf(media.mime) > -1) {
+        videoIndex = videoMimes.indexOf(media.mime) - 1
+      } else if (imageMimes.indexOf(media.mime) === -1) {
+        // unrecognized media type
+        return <div/>
+      }
+    }
+
+    if (videoIndex !== false) {
+      return (
         <video width='500' height='340' controls>
-          <source src={media.url} type={`video/${media.ext}`} />
+          <source src={assets[videoIndex].url} type={assets[videoIndex].mime}/>
           Your browser does not support the video tag.
         </video>
-      }
-    </div>)
+      )
+    }
+
+    return <ImageGrid assets={assets}/>
+  }
+
+  onFocus = (oldPost) => {
+    if (!oldPost) {
+      this.props.onFocus()
+    }
+  }
+
+  getPublishedMessage = () => {
+    const { contentSchedule } = this.props
+
+    if (!contentSchedule.id) {
+      return '[Not published - no content]'
+    }
+
+    if (contentSchedule.publishingStatus === 'PUBLISHED') {
+      return '[Published]'
+    }
+
+    if (contentSchedule.moderationStatus !== 'APPROVED') {
+      return '[Not published - not approved]'
+    }
+
+    return '[Not published]'
   }
 
   render () {
-    const {activeTab, text, medias, category, contentId, socialId, socialNetwork, comments, date, onShowCampaign, onFocus, onClick} = this.props
-    const mediaFiles = medias.map((media, index) => this.renderMedia(media, index))
+    const { contentSchedule, text, assets, moderationStatus, calendarSlot, dateTime, onShowCampaign, onFocus } = this.props
 
-    return (<Tab.Pane className={activeTab ? 'post-item-pane active' : 'post-item-pane'} key={contentId + socialNetwork}>
-      <div onClick={(e) => onClick(socialNetwork)}>
-        <div className="post-item-body">{text || 'No content'}</div>
-        {mediaFiles.length > 0 && <div className='post-item-media normal-state active'>
-          {mediaFiles.length > 1 ?
-            <div className={`image-grid col-${mediaFiles.length}`}>
-              {mediaFiles}
+    const oldPost = new Date(dateTime) < new Date()
+    let categoryColor = ''
+    if (calendarSlot !== undefined) {
+      if (calendarSlot.node.category) {
+        categoryColor = calendarSlot.node.category.color
+      }
+    }
+    let moderationButtonClass = {
+      close: '',
+      ellipsis: '',
+      check: ''
+    }
+    if (!oldPost) {
+      if (moderationStatus === 'ACCEPTED') {
+        moderationButtonClass.check = 'selected'
+      } else if (moderationStatus === 'WAITING') {
+        moderationButtonClass.ellipsis = 'selected'
+      } else if (moderationStatus === 'CHANGES_REQUIRED') {
+        moderationButtonClass.close = 'selected'
+      }
+    }
+
+    return (
+      <Tab.Pane className='post-item-pane ignore-react-onclickoutside' key={this.props._key}>
+        <div onClick={(e) => this.onFocus(oldPost)} style={{cursor: 'pointer'}}>
+          <div className='post-item-body'>{text || 'Write some content...'}</div>
+          {assets.length > 0 && <div className='post-item-media normal-state active'>
+            <div className={'image-grid' + (assets.length === 1 ? ' one' : '') + 'content-media'}>
+              {this.renderAllMedia(assets)}
             </div>
-            :
-            <div className={`image-grid col-1`}>
-              {mediaFiles}
-            </div>
+          </div>
           }
-          </div>
-        }
-      </div>
-      <div className='post-item-actions'>
-        {
-          date < new Date() &&
-          <div className='post-short-stat'>
-            <div className='social-stat'>
-              <span><Icon name='heart' color='#E8363B' />10</span>
-              <span><Icon name='retweet' color='#3B7ADB' />10</span>
-              <span><Icon name='comment' color='#939191' />10</span>
+        </div>
+        <div className='post-item-actions'>
+          <div className='left-actions'>
+            {oldPost && contentSchedule.id && contentSchedule.publishingStatus === 'PUBLISHED' &&
+            <div className='post-short-stat'>
+              <div className='social-stat'>
+                <span><Icon name='heart' />{contentSchedule.status ? contentSchedule.status.favouriteCount: 0 }</span>
+                <span><Icon name='retweet' />{contentSchedule.status ? contentSchedule.status.retweetCount : 0}</span>
+                {/* <span><Icon name='comment' />{contentSchedule.status && contentSchedule.status.retweetCount}</span> */}
+              </div>
             </div>
-          </div>
-        }
-        <Button className='category-button' style={{backgroundColor: !category.color ? '' : category.color}} onClick={() => onShowCampaign(category.id)}>
-          {category.name}
-        </Button>
-        <div className='actions'>
-          { contentId &&
+            }
+            {calendarSlot !== undefined && calendarSlot.node.category &&
+            <Button className='category-button' style={{backgroundColor: hexToRGB(categoryColor, 0.25), color: categoryColor}} onClick={() => onShowCampaign(calendarSlot.node.category.id)}>
+              {calendarSlot.node.category.name}
+            </Button>
+            }
+            {!oldPost && contentSchedule.id &&
+            <Popup content='Delete' size='tiny' trigger={
+              <span className='remove-action' onClick={() => this.props.onDeletePost()}>
+                <Icon name='trash'/>
+              </span>
+            }/>
+            }
+            {contentSchedule && contentSchedule.content &&
             <Popup
               trigger={
-                <Button className='comments-btn warning'>
-                  Feedback Recieved
-                  {/* <Label size='mini' floating circular>{comments.length}</Label> */}
-                </Button>
+                <span className='remove-action'>
+                  <Icon name='comments' />
+                </span>
               }
-              key={contentId}
+              key={contentSchedule.content.id}
               on='click'
-              open={comments && this.state.showComments}
+              open={contentSchedule.content.comments && this.state.showComments}
               onClose={this.onToggleComment}
               onOpen={this.onToggleComment}
               position='bottom right'
@@ -79,16 +140,67 @@ class NormalState extends React.Component {
             >
               <Popup.Header>Comments</Popup.Header>
               <Popup.Content>
-                <ContentComments comments={comments} contentId={contentId} socialId={socialId} />
+                <ContentComments comments={contentSchedule.content.comments.edges} contentId={contentSchedule.content.id} socialId={contentSchedule.socialProfile.id} />
               </Popup.Content>
             </Popup>
-          }
-          <span className='status success bold'>Posted</span>
-          <Logo width='14px' fill='#4ADB91' className='logo' viewBox='0 0 50 62' />
+            }
+            {!oldPost &&
+            <Popup content='Edit media' size='tiny' trigger={
+              <span className='add-image-action' onClick={() => this.props.onFocus(true)}>
+                <Icon.Group>
+                  <Icon name={'image'} inverted/>
+                  <Icon corner name={'add'} inverted/>
+                </Icon.Group>
+              </span>
+            }/>
+            }
+          </div>
+          <div className='actions'>
+            {!oldPost &&
+            <div className='approve-actions'>
+              <Popup
+                trigger={
+                  <span onClick={() => this.props.onPersist({moderationStatus: 'CHANGES_REQUIRED'})}>
+                    <Icon name='close' className={moderationButtonClass.close} />
+                  </span>
+                }
+                content='Needs Improving'
+                size='tiny'
+                position='top center'
+                inverted
+              />
+              <Popup
+                trigger={
+                  <span
+                    onClick={() => this.props.onPersist({moderationStatus: 'WAITING'})}>
+                    <Icon name='ellipsis horizontal' className={moderationButtonClass.ellipsis} />
+                  </span>
+                }
+                content='Awaiting approval'
+                size='tiny'
+                position='top center'
+                inverted
+              />
+              <Popup
+                trigger={
+                  <span
+                    onClick={() => this.props.onPersist({moderationStatus: 'ACCEPTED'})}>
+                    <Icon name='check' className={moderationButtonClass.check} />
+                  </span>
+                }
+                content='Approved'
+                size='tiny'
+                position='top center'
+                inverted
+              />
+            </div>
+            }
+            {oldPost &&
+              <small><i>{this.getPublishedMessage()}</i></small>
+            }
+          </div>
         </div>
-      </div>
-    </Tab.Pane>)
+      </Tab.Pane>
+    )
   }
 }
-
-export default NormalState
