@@ -1,15 +1,10 @@
 import React, {Component} from 'react'
-import { graphql } from 'react-apollo'
-import gql from 'graphql-tag'
-// Graphql
-import calendarSlotsGql from './graphql/calendarSlots.gql'
 
-import Loader from '../Loader'
 import PostList from './PostList'
 
 class Posts extends Component {
   render () {
-    const {calendarSlotsFind, loading, socialId, socialIds} = this.props
+    const {socialId, socialIds} = this.props
     if (!socialId) {
       return (
         <div className='page-wrapper'>
@@ -18,31 +13,12 @@ class Posts extends Component {
       )
     }
 
-    if (loading || !calendarSlotsFind) {
-      return (
-        <Loader />
-      )
-    }
-
     return (
       <div className='page-wrapper posts-page'>
-        <PostList
-          posts={calendarSlotsFind.edges}
-          socialId={socialId}
-          socialIds={socialIds}
-        />
+        <PostList socialIds={socialIds} socialId={socialId}/>
       </div>
     )
   }
 }
 
-export default graphql(gql`${calendarSlotsGql}`, {
-  options: (props) => ({
-    skip: !props.socialId,
-    variables: {
-      profileIds: [props.socialId],
-      type: 'WEEKLY'
-    }
-  }),
-  props: ({data: { calendarSlots_find, loading, error }}) => ({ calendarSlotsFind: calendarSlots_find, loading, error })
-})(Posts)
+export default Posts
