@@ -3,29 +3,35 @@ import { Helmet } from 'react-helmet'
 import { withApollo } from 'react-apollo'
 import Notifications from 'react-notify-toast'
 import PropTypes from 'prop-types'
+import { Responsive } from 'semantic-ui-react'
+
 import App from '../components/App'
 import Header from '../components/Header'
-import TopHeader from '../components/Header/TopHeader'
 import Sidebar from '../components/Sidebar'
+import MobileHeader from '../components/MobileHeader'
+import MobileSidebar from '../components/MobileSidebar'
+
 import persist from '../libraries/persist'
 
 class Auth extends Component {
   constructor (props) {
     super(props)
     this.state = {
-      sidebarOpen: true,
+      sidebarOpen: false,
       token: this.props.token,
       refreshToken: this.props.refreshToken,
       teamId: this.props.teamId,
       socialId: this.props.socialId,
-      socialIds: this.props.socialIds
+      socialIds: this.props.socialIds,
+      isOpenMobileSidebar: false,
+      isOpenMobileUserList: false
     }
   }
 
-  handleViewSidebar = () => {
-    this.setState({
-      sidebarOpen: !this.state.sidebarOpen
-    })
+  componentDidMount(){
+    if(typeof window !== undefined){
+      this.setState({sidebarOpen: (window.screen.width > 1024)? true: false})
+    }
   }
 
   useSocialProf = (socialId) => {
@@ -34,7 +40,7 @@ class Auth extends Component {
       socialId: socialId
     })
     //setTimeout(() => {
-      this.props.client.resetStore()
+      //this.props.client.resetStore()
     //}, 500)
   }
 
@@ -49,16 +55,28 @@ class Auth extends Component {
     if (typeof window !== 'undefined') window.location.reload()
   }
 
+  mobileMenuClick = () => {
+    this.setState({
+      isOpenMobileSidebar: true
+    })
+  };
+
+  mobileUserPhotoClick = () => {
+    this.setState({
+      isOpenMobileUserList: true
+    })
+  }
+
   render () {
     const { title, url, children } = this.props
-    const {sidebarOpen, token, refreshToken, socialId, socialIds, teamId} = this.state
-
+    const {sidebarOpen, token, refreshToken, socialId, socialIds, teamId, topbarData, isOpenMobileSidebar, isOpenMobileUserList } = this.state
     const childrenWithProps = React.Children.map(children,
      (child) => React.cloneElement(child, {
        token: token,
        socialId: socialId,
        socialIds: socialIds,
-       teamId: teamId
+       teamId: teamId,
+       updateTopbar: this.updateTopbar
      })
     )
 
@@ -69,25 +87,43 @@ class Auth extends Component {
           {title !== '' ? `${title} :: Proofer` : 'Proofer'}
         </title>
       </Helmet>
-      <Notifications />
-      <Header
-        token={token}
-        refreshToken={refreshToken}
-        teamId={teamId}
-        socialId={socialId}
-        socialIds={socialIds}
-        toggleSidebar={() => this.setState({ sidebarOpen: !this.state.sidebarOpen })}
-        useSocialProf={this.useSocialProf}
-        useTeamId={this.useTeamId}
-        pathname={url.pathname}
-      />
-      <Sidebar
-        activeItem={url.pathname}
-        toggleSidebar={() => this.setState({ sidebarOpen: !this.state.sidebarOpen })}
-        isOpen={this.state.sidebarOpen}
-      />
-
-      <TopHeader />
+      <Notifications options={{zIndex: 1001}} />
+      <Responsive maxWidth={Responsive.onlyMobile.maxWidth}>
+        <MobileHeader
+          socialId={socialId}
+          mobileMenuClick={this.mobileMenuClick}
+          mobileUserPhotoClick={this.mobileUserPhotoClick}
+        />
+        { isOpenMobileSidebar === true ?
+          <MobileSidebar
+            activeItem={url.pathname}
+            toggleSidebar={() => this.setState({ isOpenMobileSidebar: false })}
+            isOpen={this.state.sidebarOpen}
+            socialId={socialId}
+          /> :
+          null
+        }
+      </Responsive>
+      <Responsive minWidth={Responsive.onlyTablet.minWidth}>
+        <Header
+          token={token}
+          refreshToken={refreshToken}
+          teamId={teamId}
+          socialId={socialId}
+          socialIds={socialIds}
+          toggleSidebar={() => this.setState({ sidebarOpen: !this.state.sidebarOpen })}
+          useSocialProf={this.useSocialProf}
+          useTeamId={this.useTeamId}
+          pathname={url.pathname}
+          client={this.props.client}
+        />
+        <Sidebar
+          activeItem={url.pathname}
+          toggleSidebar={() => this.setState({ sidebarOpen: !this.state.sidebarOpen })}
+          isOpen={this.state.sidebarOpen}
+          socialId={socialId}
+        />
+      </Responsive>
       {childrenWithProps}
     </div>
     </App>)
